replace AbstractBaseChatModelIT with version using ChatModelCapabilities
