+++ langchain4j-azure-open-ai/src/test/java/dev/langchain4j/model/azure/common/AzureOpenAiChatModelIT.java
@@ -4,12 +4,12 @@ import static dev.langchain4j.model.chat.common.ChatModelCapabilities.SupportSta
 
 import dev.langchain4j.model.azure.AzureOpenAiChatModel;
 import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.chat.common.AbstractChatModelIT2;
+import dev.langchain4j.model.chat.common.AbstractChatModelIT;
 import dev.langchain4j.model.chat.common.ChatLanguageModelCapabilities;
 import dev.langchain4j.model.chat.common.ChatModelCapabilities;
 import java.util.List;
 
-class AzureOpenAiChatModelIT extends AbstractChatModelIT2 {
+class AzureOpenAiChatModelIT extends AbstractChatModelIT {
 
     // TODO https://github.com/langchain4j/langchain4j/issues/2219
