+++ langchain4j-vertex-ai/src/test/java/dev/langchain4j/model/vertexai/common/VertexAiChatModelIT.java
@@ -3,13 +3,13 @@ package dev.langchain4j.model.vertexai.common;
 import static dev.langchain4j.model.chat.common.ChatModelCapabilities.SupportStatus.NOT_SUPPORTED;
 
 import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.chat.common.AbstractChatModelIT2;
+import dev.langchain4j.model.chat.common.AbstractChatModelIT;
 import dev.langchain4j.model.chat.common.ChatLanguageModelCapabilities;
 import dev.langchain4j.model.chat.common.ChatModelCapabilities;
 import dev.langchain4j.model.vertexai.VertexAiChatModel;
 import java.util.List;
 
-class VertexAiChatModelIT extends AbstractChatModelIT2 {
+class VertexAiChatModelIT extends AbstractChatModelIT {
 
     // TODO https://github.com/langchain4j/langchain4j/issues/2219
